requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
asgiref==3.2.10           # via django
certifi==2020.6.20        # via requests, sentry-sdk
cffi==1.14.2              # via pynacl
chardet==3.0.4            # via requests
defusedxml==0.6.0         # via odfpy
diff-match-patch==20200713  # via django-import-export
django-import-export==2.3.0  # via -r requirements.in
django-localflavor==3.0.1  # via -r requirements.in
django-phonenumber-field==5.0.0  # via -r requirements.in
django==3.1               # via -r requirements.in, django-import-export, django-localflavor, django-phonenumber-field
et-xmlfile==1.0.1         # via openpyxl
gunicorn==20.0.4          # via -r requirements.in
idna==2.10                # via requests
jdcal==1.4.1              # via openpyxl
markuppy==1.14            # via tablib
odfpy==1.4.1              # via tablib
openpyxl==3.0.5           # via tablib
phonenumberslite==8.12.8  # via -r requirements.in
psycopg2==2.8.5           # via -r requirements.in
pycparser==2.20           # via cffi
pynacl==1.4.0             # via telnyx
python-stdnum==1.14       # via django-localflavor
pytz==2020.1              # via django
pyyaml==5.3.1             # via tablib
requests==2.24.0          # via -r requirements.in, telnyx
sentry-sdk==0.17.1        # via -r requirements.in
six==1.15.0               # via pynacl
sqlparse==0.3.1           # via django
tablib[html,ods,xls,xlsx,yaml]==2.0.0  # via django-import-export
telnyx==1.0.0b5           # via -r requirements.in
urllib3==1.25.10          # via requests, sentry-sdk
xlrd==1.2.0               # via tablib
xlwt==1.3.0               # via tablib

# The following packages are considered to be unsafe in a requirements file:
# setuptools
